refactor(llm): extract ChatHistoryService into services/chat-history.ts
Move getChatHistoriesById out of PluginLLM into ChatHistoryService.
Plugin call sites use chatHistory.getById(...) now. Implementation is
unchanged: same row query, groupAndTrimHistory pass, ChatMessage shape.

Via Claude Code
